Don't save game when sorting concealed tiles
A player sorting their concealed or hidden gong tiles doesn't affect the
other players, so those changes are not persisted until another
"significant" change is made. Making this private sorting be independent
should reduce some race conditions that happen when all the players are
sorting at the same time.

The ultimate version of this would be to just persist their sorted tiles
in JavaScript, but I'm not ready for that tradeoff; we're already
sacrificing complexity for performance enough as it is.