Prevent dangling peektiles when discarding
If a peektile is on the screen and a user discards from elsewhere in
their hand, it was resulting in some weird behavior. The correct
behavior is to roll the peektile into their concealed tiles.

This happens when declaring a win from their hand too. Technically this
is an illegal move, but the game does not yet have logic to check that
players go out on the most recently acquired tile.