Use Phoenix append mode for discards
This is more efficient because it doesn't transfer all discards
across the wire on every update to discards. However, it has one
drawback: if a user removes a tile from discards and then the same tile
gets put back later, it will wig out.

Realistically, this going to happen occasionally so this feature can't
be shipped. And it's not like the discards, which are plain text, are
really going to significantly slow things down anyway. I'll leave this
around in a branch as a learning exercise.

Resources:
- https://hexdocs.pm/phoenix_live_view/dom-patching.html
- https://www.youtube.com/watch?v=_ZZ1kDVwVwY
- https://newaperio.com/blog/updating-and-deleting-from-temporary-assigns-in-phoenix-live-view